+++ src/services/ai/providers/implementations.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
-import config from '../../config';
-import { logger } from '../../config/logger';
+import config from '../../../config';
+import { logger } from '../../../config/logger';
 import {
   AIProvider,
   AIGenerationOptions,